% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{active_rs_doc_copy}
\alias{active_rs_doc_copy}
\title{Copy the active document to the same location}
\usage{
active_rs_doc_copy(new = NULL, ..., old = NULL)
}
\arguments{
\item{new}{The new file name, that will be copied in the same
directory as the \link[=active_rs_doc]{active document}
For \code{active_rs_doc_move()}, a directory.}

\item{...}{These dots are for future extensions and must be empty.}

\item{old}{The old name, defaults to the active document.}
}
\value{
The new file name
}
\description{
The goal is to provide things that RStudio or usethis doesn't provide natively.
}
\details{
For example, \code{active_rs_doc_rename()} will not happen, because it is already easy
to do so via the RStudio IDE.
}
\seealso{
\code{\link[=rename_files2]{rename_files2()}}

Other document manipulation helpers:
\code{\link[=active_rs_doc_delete]{active_rs_doc_delete()}}
}
\concept{document manipulation helpers}
