% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{active_rs_doc_move}
\alias{active_rs_doc_move}
\title{Move the active document to another directory}
\usage{
active_rs_doc_move(new = NULL, old = NULL, ...)
}
\arguments{
\item{new}{A new directory}

\item{old}{The old file (defaults to the active RStudio document.)}

\item{...}{
  Arguments passed on to \code{\link{rename_files2}}
  \describe{
    \item{\code{overwrite}}{whether to overwrite \code{new} if it already exists. Be careful.}
    \item{\code{force}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}  Use \code{warn_conflicts} instead of \code{force = TRUE}}
    \item{\code{action}}{One of \code{"rename"} or \code{"test"}}
    \item{\code{warn_conflicts}}{One of
\itemize{
\item \code{"default"}: will be check more thoroughly depending on the situation. If only moving directory, and \code{"all"} otherwise.
\item \code{"all"} (larger scope: if \code{old = "data/my-streets.csv|my_streets"} will check for objects named \code{my_streets}, other files like \code{my-streets.R}, etc.),
\item \code{"exact"} will only search for \code{"data/my-streets.csv"} in documents
\code{"none"} will not search for references in documents and will rename.
}}
  }}
}
\value{
\code{new} if renaming succeeded. Mostly called for its side-effects
}
\description{
Wrapper around \code{\link[=rename_files2]{rename_files2()}}, but shortcut to allow renaming the active file.
}
