% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{active_rs_doc_nav}
\alias{active_rs_doc_nav}
\title{Open Files Pane at current document location}
\usage{
active_rs_doc_nav(path = active_rs_doc())
}
\arguments{
\item{path}{A path to file to navigate to (default active document).}
}
\value{
NULL, called for its side effects.
}
\description{
Easily navigate to active file document.
}
\details{
Wrapper around \link[rstudioapi:executeCommand]{executeCommand("activateFiles")} +
\code{\link[rstudioapi:filesPaneNavigate]{rstudioapi::filesPaneNavigate()}} + \code{\link[rstudioapi:getActiveDocumentContext]{rstudioapi::getActiveDocumentContext()}}
}
