% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-if-any.R
\name{categorize}
\alias{categorize}
\title{Categorize a vector of free text}
\usage{
categorize(variable, ..., sep = ";", ignore_case = TRUE)
}
\arguments{
\item{variable}{a character vector}

\item{...}{dynamic dots, list of recoding}
}
\value{
A named character vector
}
\description{
Related to \code{\link[=case_if_any]{case_if_any()}} but simpler!
}
\details{
It works with the same syntax as \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}
}
\examples{
categorize(
c("banana apple", "apple orange pear", "apple", "cucumber"),
 "yellow fruits" = c("banana", "pear"),
 "bad fruits" = c("orange")
)
}
