% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-files.R
\name{check_files_exist_in_dir}
\alias{check_files_exist_in_dir}
\title{Check if files referenced in source files exist in a current dir}
\usage{
check_files_exist_in_dir(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{a directory to search for}

\item{quiet}{Whether it should print messages?}
}
\value{
A logical
}
\description{
\enumerate{
\item It goes through the source files (.R/.qmd etc.),
\item It identifies data files (.csv, .xlsx) read or written
\item Search on the system if these files exist.
}
}
\keyword{internal}
