% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-files.R
\name{check_referenced_files}
\alias{check_referenced_files}
\title{Check if files referenced in source files exist in a current dir}
\usage{
check_referenced_files(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{a directory to search for}

\item{quiet}{Whether it should print messages?}
}
\value{
A logical
}
\description{
\enumerate{
\item It goes through the source files (.R/.qmd etc.),
\item It identifies data files (.csv, .xlsx) read or written
\item Search on the system if these files exist.
}
}
\details{
Still WIP, so you can add code for false positive as needed.

To find genuine referenced files, we exclude different paths
\enumerate{
\item Those created with \code{fs::path()} or \code{file.path()} or \code{glue::glue()}
\item Those that are checked for \code{fs::file_exists()}, \code{file.exists()}
\item Deleted with \code{fs::file_delete()}, \code{unlink()}
}
}
\keyword{internal}
