% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{complete_todo}
\alias{complete_todo}
\title{Remove a TODO/WORK/FIXME item from a file}
\usage{
complete_todo(line, file, regexp, rm_line = NULL)
}
\arguments{
\item{line}{The line number (a single integer)}

\item{file}{Path to a file}

\item{regexp}{A regexp to assess that the file content has not changed.}

\item{rm_line}{A logical If \code{NULL} will remove the full line in the file
(for TODO, or FIXME items), else for WORK, will only remove the WORK tag
will remove only the tag (i.e. TODO, WORK, FIXME)}
}
\value{
Writes a file with corrections, and returns the new line
content invisibly.
}
\description{
Function meant to be wrapped as \code{{.run }} hyperlinks with \code{\link[=file_outline]{file_outline()}}.
It basically removes a line from a file.
}
\keyword{internal}
