% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-plus.R
\name{count_pct}
\alias{count_pct}
\title{Count observations by group and compute percentage}
\usage{
count_pct(
  .data,
  ...,
  label = FALSE,
  accuracy = NULL,
  name = NULL,
  sort = FALSE
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to group
by.}

\item{label}{A logical, If \code{TRUE}, will return the vector as character}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}
}
\value{
An object of the same type as \code{.data}. \code{count()} and \code{add_count()}
group transiently, so the output has the same groups as the input.
}
\description{
\code{count_pct()} lets you quickly count the unique values of one or more
variables: \code{df |> count_pct(a, b)} It calculates the percentage by group
afterwards
}
\details{
Wrapper function around \code{\link[dplyr:count]{dplyr::count()}}
}
\examples{
count_pct(mtcars, cyl)
mtcars |>
  dplyr::group_by(vs) |>
  count_pct(cyl)

mtcars |>
  dplyr::group_by(vs) |>
  count_pct(cyl, label = TRUE)

mtcars |>
  dplyr::group_by(vs) |>
  count_pct(cyl, label = TRUE, accuracy = 0.1)
}
\seealso{
Other dplyr extensions:
\code{\link[=slice_min_max]{slice_min_max()}}
}
\concept{dplyr extensions}
