% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape-inline-markup.R
\name{escape_markup}
\alias{escape_markup}
\title{escape inline markup in case of problems}
\usage{
escape_markup(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with inline markup scrubbed.
}
\description{
escape inline markup in case of problems
}
\examples{
escape_markup("{gt}")
escape_markup("{.file {here}}")
escape_markup("multi problems {{gt}} to {gt} to {.file gt} to {.file {gt}}")
}
\seealso{
Other inline markup internal helpers: 
\code{\link{link_issue}()}
}
\concept{inline markup internal helpers}
\keyword{internal}
