% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{f_test}
\alias{f_test}
\title{test}
\usage{
f_test(x, y = c("a", "b", "c"), ...)
}
\arguments{
\item{x}{a value}

\item{order}{another}
}
\value{
a return
}
\description{
test
}
\examples{
f_test(x = "1", y = "c", z = "z")
}
\keyword{internal}
