% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{file_move_temp_auto}
\alias{file_move_temp_auto}
\title{Move temporary file automatically from the R console}
\usage{
file_move_temp_auto(destdir)
}
\arguments{
\item{destdir}{The desired directory to send this to}
}
\description{
It works well when you have no API to download a file, but still want a fast R implementation.
}
\seealso{
\code{\link[=file_rename_auto]{file_rename_auto()}}
}
