% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{file_rename_auto}
\alias{file_rename_auto}
\alias{file_move_dir_auto}
\title{Move file automatically between folders}
\usage{
file_rename_auto(new_name, old_file = .Last.value)

file_move_dir_auto(new_dir, old_file = .Last.value)
}
\arguments{
\item{new_name, new_dir}{New directory or file name (without extension)}

\item{old_file}{The old file name}
}
\value{
The new full path name, invisibly, allowing you to call the functions another time.
}
\description{
\itemize{
\item \code{file_rename_auto()} automatically renames your file to a better name while keeping the same folder structure
\item \code{file_move_dir_auto()} automatically moves your file while keeping the same file name
}
}
\section{Advantages}{
Instead of calling \code{fs::file_move("path/to/dir/file.R", "path/to/dir/new-file.R")}, you can just call
\code{file_rename_auto("new-file", "path/to/dir/file.R")}

Instead of calling \code{fs::file_move("path/to/dir/file.R", "path/to/new-dir/file.R")}, you can just call
\code{file_move_auto("new-dir", "path/to/dir/file.R")}

If the functions are used in conjunction with \code{\link[=file_move_temp_auto]{file_move_temp_auto()}},
}

