% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-plus.R
\name{filter_or}
\alias{filter_or}
\title{Keep rows that match one of the conditions}
\usage{
filter_or(.data, ..., .keep = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Expressions that
return a logical value, and are defined in terms of the variables in
\code{.data}. If multiple expressions are included, they are combined with the
\code{|} operator. Only rows for which \strong{one of the conditions} evaluate to \code{TRUE} are
kept.}

\item{.keep}{If \code{NULL}, will keep variables if they are named, drop otherwise.
if \code{TRUE}, will drop, if \code{FALSE}, will keep (still don't know how to implement.)}
}
\value{
An object of the same type as \code{.data}. The output has the following properties:
\itemize{
\item Rows are a subset of the input, but appear in the same order.
\item Columns are not modified (if \code{.keep = FALSE}.
\item The number of groups may be reduced (if \code{.preserve} is not \code{TRUE}).
\item Data frame attributes are preserved.
}
}
\description{
The \code{filter_or()} function is used to subset a data frame, retaining all rows
that satisfy \strong{at least one of} your conditions.
To be retained, the row must produce a value of \code{TRUE} for \strong{one of the conditions}
Note that when a condition evaluates to \code{NA} the row will be dropped, (hence this function) unlike base subsetting with \code{[}.
}
\details{
The reason to be of this function is to simplify a call like

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# with dplyr::filter
dat \%>\% dplyr::filter(vs == 1 | is.na(vs))
dat \%>\% filter_or(vs == 1, is.na(vs))
}\if{html}{\out{</div>}}

Basically, this is just a shortcut to \code{mutate(.data, new_lgl_vars)} + \code{filter(if_any(new_lgl_vars))} + \code{select(-new_lgl_vars)}
}
\examples{
library(magrittr)
mtcars \%>\% dplyr::filter(cyl > 5 | mpg == 2)
mtcars \%>\% dplyr::filter(!(!cyl > 5 & !mpg == 2))
mtcars \%>\% filter_or(cyl > 5, vs == 0)
}
\keyword{internal}
