% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape-inline-markup.R
\name{is_markup_okay}
\alias{is_markup_okay}
\title{Is inline markup valid?}
\usage{
is_markup_okay(x)
}
\value{
A logical vector
}
\description{
Is inline markup valid?
}
\examples{
is_markup_okay("{{gt}}")
is_markup_okay("{gt}")
is_markup_okay("{.file {gt}}")
}
\keyword{internal}
