% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-elements.R
\name{link_issue}
\alias{link_issue}
\title{Create a cli link to a GitHub issue}
\usage{
link_issue(x)
}
\arguments{
\item{x}{A string, usually lines of files that contains issue numbers.}
}
\value{
A formatted linked issue to GitHub issue
}
\description{
In RStudio, links to issues are automatically recognized.
This function creates cli ansi links to entries of the form rstudio/rstudio#1100
}
\details{
Note: doesn't support without \if{html}{\out{<OWNER>}}/\if{html}{\out{<REPO>}}
}
\examples{
link_issue(c("We really need rstudio/gt#1469 to be fixed.")) |> cli::cli_bullets()
}
\seealso{
Other inline markup internal helpers: 
\code{\link{escape_markup}()}
}
\concept{inline markup internal helpers}
\keyword{internal}
