% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup.R
\name{markup_href}
\alias{markup_href}
\title{Create a cli href with a markdown link}
\usage{
markup_href(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with substrings changed
}
\description{
Transforms \verb{[text](url)} -> \code{{.href [text](url)}}
}
\examples{
markup_href(c("[link](https://google.com)", "{.href [link](https://google.com)}"))

}
\seealso{
Other inline markup internal helpers:
\code{\link[=link_gh_issue]{link_gh_issue()}}
}
\concept{inline markup internal helpers}
\keyword{internal}
