% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named-base}
\alias{named-base}
\alias{min_named}
\alias{max_named}
\alias{unique_named}
\title{Helpers that can return a named vector}
\usage{
min_named(x, na.rm = FALSE, all_matches = FALSE)

max_named(x, na.rm = FALSE, all_matches = FALSE)

unique_named(x)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{Should remove \code{NA}?}

\item{all_matches}{If \code{FALSE} (default), will only return the first match,
similar to \code{\link[=which.min]{which.min()}} / \code{\link[=which.max]{which.max()}}
If \code{TRUE}, will return a named vector of all values
corresponding to the max value.}
}
\value{
\itemize{
\item \code{min/max_named(all_matches = FALSE)}, a named vector of length 1.
\item Otherwise, a named vector.
}
}
\description{
Base R keeps names in various places, but drops them elsewhere
These functions are some that I use frequently, like \code{max}, or \code{unique}
}
\examples{
max_named(c("this guy" = 2, "that guy" = 3))

unique_named(c("this guy" = 2, "that guy" = 3, "this guy" = 2))
# returns the same as base R for unnamed input
unique_named(c(1, 2, 3, 3))
# returns all values
min_named(c("x" = 1, "y" = 1, "ww" = 2), all_matches = FALSE)

# TODO is usable with `extract_cell_value()`

}
