% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdated-pkgs.R
\name{outdated_pkgs}
\alias{outdated_pkgs}
\title{Looks for outdated packages}
\usage{
outdated_pkgs()
}
\value{
A list of packages that can be updated, with links to news, the pkgdown site.
}
\description{
Only checks for binaries, which has the advantage of not notifying you when
packages are newly available on CRAN, but without the binaries available.
Installing packages from source can be longer, hence the advantage of
waiting a few days for binaries.

It takes advantage of pak's capacities to allow you to install packages on
Windows without restarting session.
}
\examples{
\donttest{
outdated_pkgs()
}
}
