% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R
\name{proj_switch}
\alias{proj_switch}
\title{Opens a RStudio project in a new session}
\usage{
proj_switch(proj = NA, new_session = TRUE)
}
\arguments{
\item{proj}{the name of a project located in the default locations or \code{NA}}

\item{new_session}{Should open in a new session?}
}
\value{
Single logical value indicating if current session is modified.
}
\description{
If not specified, will generate hyperlinks that call \code{\link[usethis:proj_activate]{usethis::proj_activate()}}.
\code{proj_switch()} looks at \code{options(reuseme.reposdir)}.
}
\seealso{
\code{\link[usethis:proj_activate]{usethis::proj_activate()}}

Other project management helpers:
\code{\link{proj-reuseme}},
\code{\link[=proj_file]{proj_file()}},
\code{\link[=proj_path]{proj_path()}}
}
\concept{project management helpers}
