% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-help.R
\name{quarto_help}
\alias{quarto_help}
\title{Show links to Quarto documentation of interest}
\usage{
quarto_help(subject = NULL)
}
\arguments{
\item{subject}{A character vector (optional)}
}
\value{
All possible links for help invisibly
}
\description{
Very opinionated of links I need to access periodically. Easily
accessible from R console.
}
\examples{
\dontshow{if (rlang::is_installed("gt")) withAutoprint(\{ # examplesIf}
gt_help <- quarto_help() |>
  tibble::enframe() |>
  gt::gt() |>
  gt::fmt_url(value)
\dontshow{\}) # examplesIf}
\dontshow{if (rlang::is_installed("pkgdown (>= 2.1.0)")) withAutoprint(\{ # examplesIf}
gt_help
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \href{https://github.com/lkwhite/Quoncierge}{Quoncierge} for an automated Quarto initialization
\item \href{https://github.com/kyleGrealis/froggeR}{froggeR} to automate Quarto
}
}
