% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Save the current image in clipboard to png in your active directory}
\usage{
screenshot(file = NULL, dir = NULL)
}
\arguments{
\item{file}{A file name, ideally \code{-} (kebab-case). (extension ignored) (optional, default is \code{image.png})}

\item{dir}{A directory (optional), to override the directory rules mentioned in the description.}
}
\value{
The full image path, invisibly.
}
\description{
The screenshot will be saved as \code{.png} to a directory following these rules
\enumerate{
\item In a regular RStudio project (or a Quarto book), it will be saved to a \verb{images/} directory
\item In a package project, it will be saved in a \code{man/figures} directory
\item In a Quarto Blog project, it will save in the current post's folder.
\item You can always override these defaults by setting \code{dir}
}

After using the shortcut Win + Shift + S, you can call this function!
}
\details{
If no file name is supplied, a file named \code{image0*.png} will be created.
The function then prompts you to rename the file with a more expressive name.
It will continue the numbering if a file named image exists.

Still have to validate if it works on macOS, as it is not clear whether the image goes to the clipboard by default
}
\examples{
if (FALSE) {
  # Add an image to the clipboard
  # Run the following
  screenshot(file = "my-new-image")
}

}
