% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-check.R
\name{select_check}
\alias{select_check}
\title{Give a more informative error in case of tidyselect errors}
\usage{
select_check(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}
}
\description{
Give a more informative error in case of tidyselect errors
}
\examples{
try(mtcars |> select_check(vs3))
}
