% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-plus.R
\name{slice_group_sample}
\alias{slice_group_sample}
\title{Explore all rows in a random group}
\usage{
slice_group_sample(data, n_groups = 1, group_var = NULL)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{n_groups}{Number of groups to sample. (passed to \code{sample(size = n_groups)})}

\item{group_var}{if the data is grouped, will be ignored.}
}
\value{
A data frame with a sample group.
}
\description{
Compared to \code{slice_sample()} \code{slice_group_sample} will return all rows corresponding to a group.
}
\examples{
set.seed(10)
library(magrittr)
slice_group_sample(mtcars, group_var = vs)
mtcars \%>\%
  dplyr::group_by(vs) \%>\%
  slice_group_sample()
}
\concept{family dplyr extensions}
