% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-conflicts.R
\name{solve_file_name_conflict}
\alias{solve_file_name_conflict}
\title{Check if outdated or non-existent file is.}
\usage{
solve_file_name_conflict(
  files,
  regex,
  dir = ".",
  extra_msg = NULL,
  quiet = FALSE,
  what = NULL
)
}
\arguments{
\item{files}{which files to search in}

\item{regex}{a regex related to the file name to search for}

\item{dir}{A directory where to operate}

\item{extra_msg}{Extra message to pass}

\item{quiet}{A logical, informs where the occurrences are found. (Default, \code{FALSE})}

\item{what}{Which file conflicts we talking about}
}
\value{
Mostly called for its side-effects, but will return the number of matches
(0 if no referenced files are problmatic)
}
\description{
If \code{quiet = FALSE} (default) will give a hint of where the file is referenced.
}
\keyword{internal}
