% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-df.R
\name{visual_diff_df}
\alias{visual_diff_df}
\title{Get a widget that displays diff between data frame}
\usage{
visual_diff_df(old, new, width = NULL, height = NULL)
}
\arguments{
\item{old, new}{Pair of data frames to compare}

\item{width, height}{Output size}
}
\value{
A HTML widget
}
\description{
Thin wrapper around \code{\link[diffviewer:visual_diff]{diffviewer::visual_diff()}}.
}
\details{
Also writes to tempfiles to make sure it works
}
\examples{

mtcars2 <- mtcars

mtcars2[1, 1] <- NA
visual_diff_df(old = mtcars, new = mtcars2)
}
