% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-excel.R
\name{write_temp_excel}
\alias{write_temp_excel}
\title{Write temporary Excel to Downloads}
\usage{
write_temp_excel(dat, name, font = "Arial", na = "", dir = NULL, mip = NULL)
}
\arguments{
\item{dat}{data frame}

\item{name}{name of file}

\item{font}{font name}

\item{na}{passed to \code{\link[openxlsx2:wb_add_data]{openxlsx2::wb_add_data()}}}

\item{dir}{A directory to override.}

\item{mip}{privacy string}
}
\value{
A new Excel file
}
\description{
Write temporary Excel to Downloads
}
