% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{active_rs_doc_delete}
\alias{active_rs_doc_delete}
\title{Delete the active RStudio document safely}
\usage{
active_rs_doc_delete()
}
\value{
Called for side-effects. The document content invisibly if deleting and reason.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gathers informative summary about the document you are about to delete.

Will delete more easily if file name starts with \verb{temp-}, if file is untracked and recent.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
active_rs_doc_delete()
\dontshow{\}) # examplesIf}
}
\seealso{
Other document manipulation helpers: 
\code{\link{active_rs_doc_copy}()}
}
\concept{document manipulation helpers}
