% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-if-any.R
\name{categorize}
\alias{categorize}
\title{Categorize a vector of free text}
\usage{
categorize(variable, categories, sep = ";", ignore_case = TRUE)
}
\arguments{
\item{categories}{a list}
}
\description{
Categorize a vector of free text
}
\examples{
categorize(
c("banana apple", "apple orange pear", "apple", "cucumber"),
list(
  "yellow fruits" = c("banana", "pear"),
  "bad fruits" = c("orange")
)
)
}
