% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named-base}
\alias{named-base}
\alias{max_named}
\alias{min_named}
\alias{unique_named}
\title{Helpers that can return a named vector}
\usage{
max_named(x, na.rm = FALSE)

min_named(x, na.rm = FALSE)

unique_named(x)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{Should remove \code{NA}}
}
\value{
A vector (preserves names)
}
\description{
Base R keeps names in various places, but drops them elsewhere
These functions are some that I use frequently, like \code{max}, or \code{unique}
}
\examples{
max_named(c("this guy" = 2, "that guy" = 3))

unique_named(c("this guy" = 2, "that guy" = 3, "this guy" = 2))
# returns the same as base R for unnamed input
unique_named(c(1, 2, 3, 3))
}
