% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_rs_doc}
\alias{open_rs_doc}
\alias{active_rs_doc}
\alias{active_rs_dir}
\title{Open a Document in RStudio}
\usage{
open_rs_doc(path, line = -1L, col = -1L, move_cursor = TRUE)

active_rs_doc()

active_rs_dir()
}
\arguments{
\item{path}{The path to the document.}

\item{line}{The line in the document to navigate to.}

\item{col}{The column in the document to navigate to.}

\item{move_cursor}{Boolean; move the cursor to the requested location after
opening the document? s}
}
\value{
Invisibly returns the document ids
}
\description{
Wrapper around \code{\link[rstudioapi:rstudio-documents]{rstudioapi::documentOpen()}}, but with \verb{fs paths}, for consistency.
If the file could not be opened, a clickable hyperlink is displayed.
}
\details{
\itemize{
\item \code{active_rs_doc()} is a wrapper around \code{\link[rstudioapi:rstudio-documents]{rstudioapi::documentPath()}} that handles
unsaved files gracefully
\item \code{active_rs_dir()} is a shortcut for \verb{fs::path_dir(active_rs_doc()} and returns the project path
if error.
}
}
\examples{
if (FALSE) {
  # open the fictious file.R at line 5
  open_rs_doc("file.R", line = 5)
}

}
