% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_rs_doc}
\alias{open_rs_doc}
\title{Open a Document in RStudio}
\usage{
open_rs_doc(path, line = -1L, col = -1L, move_cursor = TRUE)
}
\arguments{
\item{path}{The path to the document.}

\item{line}{The line in the document to navigate to.}

\item{col}{The column in the document to navigate to.}
}
\value{
Invisibly returns the document id
}
\description{
Wrapper around \link[rstudioapi:rstudio-documents]{rstudioapi::documentOpen}
}
