% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\name{outline}
\alias{outline}
\alias{file_outline}
\alias{proj_outline}
\alias{dir_outline}
\title{Print interactive outline of file sections}
\usage{
file_outline(
  pattern = NULL,
  path = active_rs_doc(),
  work_only = TRUE,
  alpha = FALSE,
  dir_common = NULL,
  exclude_todos = FALSE,
  recent_only = FALSE,
  print_todo = deprecated()
)

proj_outline(
  pattern = NULL,
  proj = proj_get2(),
  work_only = TRUE,
  exclude_tests = FALSE,
  exclude_todos = FALSE,
  dir_tree = FALSE,
  alpha = FALSE,
  recent_only = FALSE
)

dir_outline(
  pattern = NULL,
  path = ".",
  work_only = TRUE,
  exclude_tests = FALSE,
  exclude_todos = FALSE,
  dir_tree = FALSE,
  alpha = FALSE,
  recent_only = FALSE,
  recurse = FALSE
)
}
\arguments{
\item{pattern}{A string or regex to search for in the outline. If
specified, will search only for elements matching this regular expression.
The print method will show the document title for context. Previously \code{regex_outline}.}

\item{path, proj}{A character vector of file paths, a \link[=proj_list]{project}.
Defaults to active file, project or directory. \code{rstudioapi::documentPath()}}

\item{work_only}{If \code{TRUE}, (the default), will only show you work items
first. Set to \code{FALSE} if you want to see the full outline. \code{WORK} will
combine with \code{pattern}}

\item{alpha}{Whether to show in alphabetical order}

\item{dir_common}{(Do not use it)}

\item{exclude_todos}{Should include TODOs in the file outline? If \code{FALSE}, will
print a less verbose output with sections.}

\item{recent_only}{Show outline for recent files}

\item{print_todo}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. Use \code{exclude_todos} instead.}

\item{exclude_tests}{Should tests be displayed? (Not sure if this argument will stay, Will have to think)}

\item{dir_tree}{If \code{TRUE}, will print the \code{\link[fs:dir_tree]{fs::dir_tree()}} or non-R files in
the directory}

\item{recurse}{If \code{TRUE} recurse fully, if a positive number the number of levels
to recurse.}
}
\value{
A \code{outline_report} object that contains the information. Inherits
\code{tbl_df}.

A symbol will show for recently modified files.
}
\description{
THe outline functions return a data frame that contains details of file location.

It also includes a print method that will provide a console output that will include \href{https://cli.r-lib.org/reference/links.html}{clickable hyperlinks}
in RStudio (or if your terminal supports it). It works with both (qR)md and R files.

Outline elements include
\itemize{
\item Any code section
\item function definition (not shown in console by default)
\item \code{TODO} items
\item Parse cli hyperlinks
\item Plot or table titles
\item FIgures caption in Quarto documents
\item test names
\item Indicator of recent modification
\item Colored output for
\item URL and gh issue detection and link creation.
}

If \code{work_only} is set to \code{TRUE}, the function will only return outline of the \verb{# WORK} comment
in \code{path}. \code{work_only = TRUE} will have an effect on \code{pattern}.

By default
\itemize{
\item \code{file_outline()} prints the outline the \link[=active_rs_doc]{active document} if in RStudio
\item \code{proj_outline()} prints the outline of the \link[usethis:proj_utils]{active project} if in RStudio
\item \code{dir_outline()} prints the outline of the \link[=getwd]{active working directory} by default or
}
}
\details{
\code{proj_outline()} and \code{dir_outline()} are wrapper of \code{file_outline()}.

The parser is very opinionated and based on lightparser, roxygen2 and regexps.

Will show TODO items and will offer a link to \link[=complete_todo]{mark them as complete}.
}
\examples{
file <- fs::path_package("reuseme", "example-file", "outline-script.R")
file_outline(path = file)

# Remove todo items
file_outline(path = file, exclude_todos = TRUE, alpha = TRUE)

# interact with data frame
file_outline(path = file) |> dplyr::as_tibble()

\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# These all work on the active file / project or directory.

file_outline()
proj_outline()
dir_outline()
# Like proj_switch(), proj_outline() accepts a project
\dontshow{\}) # examplesIf}
}
