% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-reuseme.R
\name{proj-reuseme}
\alias{proj-reuseme}
\title{Interact with different RStudio projects}
\description{
The package offers many ways to interact with different local RStudio projects.
}
\section{Setup}{
To take advantage of this functionality, you first have to set \code{options(reuseme.reposdir)} in
your .Rprofile file. Access it with \code{\link[usethis:edit]{usethis::edit_r_profile()}}.

I would recommend you add the following. It works better if you store your RStudio
projects in common directories.

Inspired by \link[usethis:usethis_options]{usethis options}

\if{html}{\out{<div class="sourceCode">}}\preformatted{if (interactive()) \{
  options(reuseme.reposdir = c("~/rrr/", "~/packages", "~/rrr-work/"))

\}
}\if{html}{\out{</div>}}
}

\section{Capabilities.}{
Assumes that you have a project named \code{"learning"}
A project outline

\if{html}{\out{<div class="sourceCode">}}\preformatted{proj_outline(proj = "learning)
}\if{html}{\out{</div>}}

Add a TODO item to the \code{learning} project

\if{html}{\out{<div class="sourceCode">}}\preformatted{use_todo("learning::Learn this")
}\if{html}{\out{</div>}}

Get file \link[=proj_outline]{outline} of the \code{file.R} in "learning"

\if{html}{\out{<div class="sourceCode">}}\preformatted{proj_file("file", "learning")
}\if{html}{\out{</div>}}

Move to a new project in the same session

\if{html}{\out{<div class="sourceCode">}}\preformatted{proj_switch("learning")
}\if{html}{\out{</div>}}

A lot of these features are already present in RStudio and with usethis.
However, when managing many projects, the recent projects list can be more difficult
to handle.
Passing the full project name to \code{usethis::proj_activate()} was too long.
}

\seealso{
Other project management helpers: 
\code{\link{proj_file}()},
\code{\link{proj_list}()},
\code{\link{proj_switch}()}
}
\concept{project management helpers}
