% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R
\name{proj_file}
\alias{proj_file}
\title{Access the file outline within other project}
\usage{
proj_file(file = NULL, path = active_rs_proj(), pattern = NULL)
}
\arguments{
\item{file}{A filename or regexp to a file inside \code{proj}}

\item{path}{a project path  \code{\link[=proj_list]{proj_list()}}. If \code{NULL},
will return active project.}

\item{pattern}{A regular expression to look for}
}
\value{
The file outline if multiple matches are found
}
\description{
It can be used as \code{\link[=file_outline]{file_outline()}} + \code{proj}.
}
\examples{
try(proj_file("A non-existent file"))
}
\seealso{
Other project management helpers: 
\code{\link{proj-reuseme}},
\code{\link{proj_list}()},
\code{\link{proj_switch}()}
}
\concept{project management helpers}
