% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R
\name{proj_file}
\alias{proj_file}
\title{Active a file at location}
\usage{
proj_file(file = NULL, proj = NULL, pattern = NULL)
}
\arguments{
\item{file}{A filename or regexp to a file inside \code{proj}}

\item{proj}{a project path or file \code{\link[=proj_list]{proj_list()}}}

\item{pattern}{A regular expression to look for}
}
\value{
The file outline if multiple matches are found
}
\description{
Active a file at location
}
\examples{
try(proj_file("A non-existent file"))
}
\seealso{
Other project management helpers: 
\code{\link{proj-reuseme}},
\code{\link{proj_list}()},
\code{\link{proj_switch}()}
}
\concept{project management helpers}
