% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R
\name{proj_list}
\alias{proj_list}
\title{Returns a named project list options}
\usage{
proj_list(proj = NULL, dirs = getOption("reuseme.reposdir"))
}
\arguments{
\item{proj}{A project path or name to match}

\item{dirs}{The directories in which we want to list projects.}
}
\value{
A named character vector with the project name as name, and path as value.
If \code{proj} is supplied
}
\description{
It peeks \code{options(reuseme.reposdir)} to find projects.
}
\seealso{
Other project management helpers: 
\code{\link{proj-reuseme}},
\code{\link{proj_file}()},
\code{\link{proj_switch}()}
}
\concept{project management helpers}
