% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R, R/proj.R
\name{proj_list}
\alias{proj_list}
\title{Returns a named project list options}
\usage{
proj_list(dirs = getOption("reuseme.reposdir"))

proj_list(dirs = getOption("reuseme.reposdir"))
}
\arguments{
\item{dirs}{The directories in which we want to list projects.}
}
\value{
A named character vector with the project name as

A named character vector with the project name as
}
\description{
It peeks \code{options(reuseme.reposdir)} to find projects.

It peeks \code{options(reuseme.reposdir)}
}
