% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-list.R
\name{proj_path}
\alias{proj_path}
\alias{proj_list}
\title{Specify \code{proj} in functions}
\usage{
proj_path(proj = NA, ..., dirs = getOption("reuseme.reposdir"))

proj_list(proj = NA, dirs = getOption("reuseme.reposdir"))
}
\arguments{
\item{proj}{A project path or name to match. If \code{NA}, returns all projects.
If \code{NULL}, returns the active project.}

\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{dirs}{The directories in which we want to list projects.}
}
\value{
A named character vector with the project name as name, and path as value.
If \code{proj} is supplied
}
\description{
\itemize{
\item \code{proj_}
Two main ways to specify proj:
\item Set \code{options(reuseme.reposdir)} in \code{.Rprofile} that contains a character
vector of paths to your projects. (i.e. \verb{~/rrr} that contains all your projects)
\item Specify the full path to \code{proj}. (like you would for usethis function)
}
}
\seealso{
Other project management helpers: 
\code{\link{proj-reuseme}},
\code{\link{proj_file}()},
\code{\link{proj_switch}()}
}
\concept{project management helpers}
