% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-files.R
\name{rename_files2}
\alias{rename_files2}
\title{Rename an output or a data file}
\usage{
rename_files2(old, new, force = FALSE, action = c("rename", "test"))
}
\arguments{
\item{old, new}{Old and new file names (with or without extensions).}

\item{force}{Whether to force renaming if there are conflicts}

\item{action}{One of \code{"rename"} or \code{"test"}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function can improve your workflow.
It is inspired by \code{\link[usethis:rename_files]{usethis::rename_files()}}, but its scope
is more oriented towards analysis script.
}
\section{Use case}{
Let's say you have an analysis and work on a certain subject.
You want to rename a figure for clarity.

Here is what \code{rename_files2()} does for you, before it renames files.
\enumerate{
\item Look for potential name conflict
\item Look for data frame name conflicts
\item Sends information to clipboard
}
}

