% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuseme-package.R
\docType{package}
\name{reuseme-package}
\alias{reuseme}
\alias{reuseme-package}
\title{reuseme: Collections of Utility Functions to Work Across Projects}
\description{
Allows you to browse current projects, rename files safely, add screenshots to project on Windows. It is also my personal library and contains wrapper around common functions, from dplyr and readxl. It takes advantage of cli hyperlinks. Finally, it provides a custom print method for tibbles, inspired by janitor, and readr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://olivroy.github.io/reuseme/}
  \item \url{https://github.com/olivroy/reuseme}
  \item Report bugs at \url{https://github.com/olivroy/reuseme/issues}
}

}
\author{
\strong{Maintainer}: Olivier Roy \email{olivierroy71@hotmail.com}

}
\keyword{internal}
