% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Save the image in your clipboard to png  in a \verb{figures/} dir}
\usage{
screenshot(file = NULL)
}
\arguments{
\item{file}{A file name, ideally \code{-} (kebab-case). (extension ignored)}
}
\value{
invisibly returns the full image path
}
\description{
Creates a \verb{figures/} directory if in RStudio Project. Will create a \code{man/figures} in a package.
After using the shortcut Win + Shift + S, you can call this function
}
\details{
Still have to validate if it works on MacOS, as it is not clear whether the image goes to the clipboard by default
}
\examples{
\dontrun{
screenshot(file = "my-new-image")
}
}
