% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Save the current image in clipboard to png in your active directory}
\usage{
screenshot(file = NULL, proj = proj_get(), dir = NULL)
}
\arguments{
\item{file}{A file name, ideally \code{-} (kebab-case). (extension ignored) (optional, default is \code{image.png})}

\item{proj}{A project name}

\item{dir}{A directory (optional), to override the directory rules mentioned in the description. inside \code{proj}.}
}
\value{
The full image path, invisibly.
}
\description{
The screenshot will be saved as \code{.png} to a directory following these rules
\enumerate{
\item In a regular RStudio project (or a Quarto book), it will be saved to a \verb{images/} directory
\item In a package project, it will be saved in a \code{man/figures} directory
\item In a Quarto Blog project, it will save in the current post's folder.
\item You can always override these defaults by setting \code{dir}
}

After using the shortcut Win + Shift + S, you can call this function!
}
\details{
If no file name is supplied, a file named \code{image0*.png} will be created.
The function then prompts you to rename the file with a more expressive name.
It will continue the numbering if a file named image exists.

Still have to validate if it works on macOS, as it is not clear whether the
image goes to the clipboard by default

The maximum number of images in a folder is 99. (only padding 2), should be enough.

You should not be able to overwrite a screenshot with a generic name, only a
named one as it is possible you may require to retake your screenshot.
}
\examples{
if (FALSE) {
  # Add an image to the clipboard
  # Run the following
  screenshot(file = "my-new-image")
}

}
