% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo-utils.R
\name{use_todo}
\alias{use_todo}
\title{Add a TODO list by project to a TODO.R file in the base directory}
\usage{
use_todo(todo, proj = proj_get(), code = FALSE)
}
\arguments{
\item{todo}{A character vector of lines to add to the TODO file}

\item{proj}{By default, the active project, an arbitrary directory, or a
RStudio project name in the following directories
\code{options(reuseme.destdir)}, uses \code{\link[=proj_list]{proj_list()}} in this case.}

\item{code}{If \code{TRUE}, will render code output (default is text.)}
}
\value{
A \code{TODO.R} file appended with the \code{todo} string.
}
\description{
Creates or edits a \code{TODO.R} file to store your TODOs.
By default it will write in the current RStudio project.
}
\details{
If you use \code{use_todo()} with a version-control repository, you may want to use
\code{usethis::use_git_ignore("TODO.R")} if you don't want your \code{TODO.R} file to be included.
If using in a package directory, use \code{usethis::use_build_ignore("TODO.R")} to prevent a note in \verb{R CMD CHECK}
}
\examples{
if (FALSE) {
  use_todo("I need to do that")
  use_todo(c("I need to do that again", "youppi"))
  use_todo("c(x, y)", code = TRUE)
}
}
\seealso{
\code{\link[usethis:write-this]{usethis::write_union()}}
}
